Backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-dotenv==1.0.1
llama-index==0.11.22
pydantic==2.9.2